gh-155. Move handlers to hendlers folder
